# openapi.yaml
openapi: 3.1.0
info:
  title: ForemanOS API
  version: 1.0.0
  summary: AI-powered construction project management API
  description: |
    AI-powered construction project management platform API.

    **Authentication:** Most endpoints require JWT authentication via NextAuth.

    **Rate Limits:**
    - Chat API: 20 requests/minute
    - Upload API: 10 requests/minute
    - Auth API: 5 requests/5 minutes
    - General API: 60 requests/minute

    **Subscription Tiers:** Free, Starter, Pro, Team, Business, Enterprise
  contact:
    name: ForemanOS Support
    url: https://foremanos.site
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://foremanos.site
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/documents/upload:
    post:
      tags:
        - Documents
      summary: Upload a document
      description: |
        Upload construction documents (PDF, images, DOCX) for processing.

        **Features:**
        - Duplicate detection via file hash
        - Automatic document classification
        - Vision-based processing for PDFs and images
        - Subscription quota enforcement
        - Async processing with status tracking

        **Rate Limit:** 10 uploads/minute per user
        **Max File Size:** 200MB
        **Allowed Types:** PDF, JPEG, PNG, TIFF, DOCX, TXT
      operationId: uploadDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - projectId
              properties:
                file:
                  type: string
                  format: binary
                  description: The document file to upload
                projectId:
                  type: string
                  format: uuid
                  description: Project ID to associate document with
                  example: "550e8400-e29b-41d4-a716-446655440000"
                category:
                  type: string
                  enum:
                    - architectural
                    - structural
                    - mep
                    - civil
                    - specifications
                    - contracts
                    - submittals
                    - rfi
                    - other
                  default: other
                  description: Document category for organization
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  Document:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      name:
                        type: string
                      fileName:
                        type: string
                  message:
                    type: string
                    example: "Document uploaded successfully. Processing in background..."
                  processingInfo:
                    type: object
                    properties:
                      estimatedPages:
                        type: integer
                        description: Estimated number of pages
                      processorType:
                        type: string
                        enum: [vision, text, hybrid]
                        description: Processing method used
                      remainingPages:
                        type: integer
                        description: Remaining pages in quota
                      tier:
                        type: string
                        enum: [free, starter, pro, team, business, enterprise]
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Permission denied or quota exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Processing quota exceeded"
                  message:
                    type: string
                  remainingPages:
                    type: integer
                  tier:
                    type: string
                  classification:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Duplicate document detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Duplicate document detected"
                  message:
                    type: string
                    example: "A document with the same name and content already exists in this project."
        '413':
          description: File too large
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "File too large. Maximum size is 200MB."
        '415':
          description: Unsupported file type
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid file type"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth JWT token from session
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Message or image is required"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"
    RateLimitExceeded:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry is allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Too many requests. Please slow down and try again in a moment."
              retryAfter:
                type: integer
                description: Seconds until user can retry
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
              details:
                type: string
                description: Additional error details (dev mode only)
...